Switch to NES 2.0 for ROM outputs
No-Intro has switched over to NES 2.0 for everything, meaning it doesn't
list hashes for the older header format. Use a NES 2.0 header when
assembling ROMs so that we match these new hashes and it's easier to
verify that this repo is building what it says it is.

Also add NES 2.0 ROM input hash for Mother 1, and remove a hardcoded
reference to /bin/bash that doesn't work on Windows.